aktualizace karet

+++ Banka.cs
@@ -10,8 +10,8 @@ namespace semestralka_windows_forms
     {
         public uint ID { get; private set; }
         public DateTime Datum { get; private set; }
-        public int Castka { get; private set; }
-        public Banka(uint id, DateTime platba, int castka)
+        public decimal Castka { get; private set; }
+        public Banka(uint id, DateTime platba, decimal castka)
         {
             ID = id;
             Datum = platba;
